requirements.txt
astroid==3.1.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
colorama==0.4.6
dill==0.3.8
h11==0.14.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
outcome==1.3.0.post0
platformdirs==4.2.0
pycparser==2.21
pylint==3.1.0
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
tomlkit==0.12.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
